+++ src/main/java/server/MATE/domain/payment/service/MockPaymentService.java
@@ -3,10 +3,10 @@ package server.MATE.domain.payment.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import server.MATE.domain.payment.dto.PaymentCreateResponse;
-import server.MATE.domain.payment.dto.PaymentExecuteResponse;
-import server.MATE.domain.payment.dto.PaymentRefundResponse;
-import server.MATE.domain.payment.dto.PaymentStatusResponse;
+import server.MATE.domain.payment.dto.response.PaymentCreateResponse;
+import server.MATE.domain.payment.dto.response.PaymentExecuteResponse;
+import server.MATE.domain.payment.dto.response.PaymentRefundResponse;
+import server.MATE.domain.payment.dto.response.PaymentStatusResponse;
 import server.MATE.domain.payment.entity.Payment;
 import server.MATE.domain.payment.entity.PaymentRefund;
 import server.MATE.domain.payment.repository.PaymentRefundRepository;
